fix build issue related to compatibility of nextauth with nextjs  app router
